Added 404 page. Made login and register more beautiful. Some changes in Post View

+++ Web/StarForums.Web/Controllers/PostController.cs
@@ -47,6 +47,11 @@
         [Authorize]
         public IActionResult Create(string categoryName)
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.Redirect("/Login");
+            }
+
             var model = new CreatePostInputModel();
 
             var test = this.categoriesService.GetByName<CategoryViewModel>(categoryName);
@@ -61,6 +66,11 @@
         [HttpPost]
         public async Task<IActionResult> Create(CreatePostInputModel input)
         {
+            if (!this.User.Identity.IsAuthenticated)
+            {
+                return this.Redirect("/Login");
+            }
+
             if (!this.ModelState.IsValid)
             {
                 return this.View(input);
@@ -79,9 +89,16 @@
         [Authorize]
         public async Task<IActionResult> Delete(int postId, string categoryName)
         {
-            await this.postsService.Delete(postId);
+            var post = this.postsService.GetById<Post>(postId);
 
-            return this.Redirect("/" + categoryName);
+            if (this.userManager.GetUserId(this.User) == post.UserId || this.User.IsInRole("Administrator"))
+            {
+                await this.postsService.Delete(postId);
+
+                return this.Redirect("/" + categoryName);
+            }
+
+            return this.NotFound();
         }
 
         [Route("/{categoryName}/{postId}/AddComment")]
